Improved Validations

+++ src/controllers/Product.controller.js
@@ -11,9 +11,14 @@ class ProductController {
         const {product_name, category, description, price, adicional_content} = req.body;
 
         const verify_category = await database.find(ProductCategoryModel, {category_title: category})
+        const getProduct = await database.find(Product, {product_name: product_name})
 
         if(verify_category.length === 0){
             return res.status(400).send({error: "Please, enter a valid category name"})
+        }
+
+        if(getProduct.length > 0) {
+            return res.status(400).send({error: "This product already exists"})
         }
 
         try {
@@ -36,16 +41,23 @@ class ProductController {
     static async get(req, res) {
         const {product_name} = req.body;
 
-        try{
-            res.status(202).send({success: await database.find(Product, {product_name: product_name})})
-        } catch (error) {
-            return res.status(400).send({error: "error in request"})
+        const getProduct = await database.find(Product, {product_name: product_name})
+
+        if(getProduct.length === 0) {
+            return res.status(400).send({error: "This product doesn't exists"})
         }
 
+        return res.status(200).send({success: getProduct})
     }
 
     static async delete(req, res) {
         const {id} = req.body;
+
+        const getProduct = await database.find(Product, {id: id})
+
+        if(getProduct.length === 0) {
+            return res.status(400).send({error: "This product doesn't exists"})
+        }
 
         try {
             await database.delete(Product, {id: id})
@@ -60,6 +72,12 @@ class ProductController {
 
     static async update(req, res) {
         const {id, product_name, category, description, price, adicional_content} = req.body;
+
+        const getProduct = await database.find(Product, {product_name: product_name})
+
+        if(getProduct.length > 0) {
+            return res.status(400).send({error: "This product already exists"})
+        }
 
         try {
             await database.update(Product, {id: id}, {
